chore(ci): Ignore removal of _cq_sync_time and _cq_source_name for SDK v4 migration (#11905)
The `_cq_sync_time` and `_cq_source_name` will now be added by the CLI as a transformation, and are not considered part of the source tables anymore. That said, this should not be considered a breaking change, as the default behavior will be to keep these columns in place.