feat(aws-policies): Add in AWS security account contact query (#11729)
Adding Foundational Security Best Practice Policy - net new policy that we haven't covered for AWS Accounts.

This check looks for the presence of a security contact for each account.  

This check tested the following use cases.  These cases cover all expected pass and fail use cases for security contacts which are contained within the `aws_account_alternate_contact` table.  These alternate tables can include operations, billing, or security:

- Account with no alternate contacts (expected fail)
- Account with only billing or operations and no security contact (expected fail)
- Account with only a security contact (expected pass)
- Account with a security contact and other contacts (expected positive)


To do so, I joined this with a table that references the AWS Account.  This is not an ideal join, but the query uses `aws_iam_accounts` to get an idea of the account we're pulling data from.  Another option would be to use ` aws_account_contacts`, but that requires assuming that accounts will always have contact information (which should generally be the case).  The other tables considered were `aws_organizations_accounts` but that may not return the data we want due to permissions within the organization and how accounts could be listed depending on visibility.  

A couple conditions - updating contact information in AWS requires all fields to be populated.  Additionally, a lack of security contact will return no data - so we do a NULL check based off a left join (with iam_accounts as the left table).  We remove the non-security contacts in the right join (nested select).  

Right now, we pull contact information from the calling AWS Account.  It's possible to pull contact information from a delegated administrator or an administrator account.  If we revisit that later, this query will need to be updated.