feat(postgresql)!: Migrate Postgres destination to SDK v4 (#11662)
Upgrades the PostgreSQL destination to plugin SDK V4. This changes the plugin to use the new streaming interface, allowing migrations to happen more than once during a sync and for PostgreSQL to be used as a state backend.

While there shouldn't be anything breaking in this upgrade, I am marking this as a breaking change because it is a big internal refactor, and despite our best efforts there may be some things that we missed during testing.