fix(mysql)!: Limit truncation of `text` and `blob` columns used as primary Keys by setting the prefix to the largest value (#11982)
#### Summary

Closes #11962 
BEGIN_COMMIT_OVERRIDE
fix!: This changes the prefix size for `text` and `blob` primary keys. Previous to this PR only the first 64 characters were included in the primary key, but now it will use maximum supported number of characters (191) in the primary key. Warning logs will be printed if there is any possible truncation. If you are using a source plugin that has a primary key on a `text` or `blob` column you will have to use `migrate_mode: forced` to upgrade.
END_COMMIT_OVERRIDE