feat: Support date types (#12538)
#### Summary

Fixes https://github.com/cloudquery/cloudquery/issues/12524

When reading `parquet` files, DuckDB seems to be failing to identify `date` types, thinks those are `timestamp without timezone` and fails since it doesn't know how to convert a timestamp to date. Found https://github.com/duckdb/duckdb/issues/1291 and other `parquet` related issues:
https://github.com/duckdb/duckdb/issues?q=is%3Aissue+is%3Aopen+parquet+
https://github.com/marcboeker/go-duckdb/issues?q=is%3Aissue+parquet+is%3Aclosed

The solution I found is to use `timestamp` to store date types when writing/reading to/from the temporary `parquet` file. It might take up a bit more space in the `parquet` file but it works.

We might want to look into skipping the temporary file if it's not a huge effort to do

<!--