chore: Exclude external links in broken links checker (#12844)
Opening this one to see what others think. I like the broken links checker, but external links seem to just lead to a never-ending stream of toil to fine-tune what we exclude, and is prone to failure when external websites are having downtime. I'd rather test what we can control, and ignore the rest.