feat: Add protocol v3 tables command (move doc generation to CLI) (#12758)
This seems to work, but marking as draft because without the change in https://github.com/cloudquery/plugin-sdk/pull/1129 the docs will be missing some information. I also still need to update the code here to populate table relationship data based on the parent fields by walking the tree, so that this information will be correctly displayed in the docs.