fix!: Update to latest Python SDK (#13051)
Upgrade to the latest Python SDK and generate docs using `cloudquery tables` command. The latest Python SDK removes the markdown docs generation capability from the SDK, so this is also no longer directly available from the plugin (use `cloudquery tables` instead).

There are no real breaking schema changes in this PR; only the names of the types are being changed to be consistent with those of the Go plugins.