fix!: Upgrade SDK to v0.1.1 (#13055)
This upgrades the Python SDK dependency to 0.1.1, and starts generating documentation using the `cloudquery tables` command instead of the (now-deprecated) plugin `doc` command.

Note: there are no breaking schema changes in this PR - only different aliases of the types are used to be in line with those used by Go plugins. The breaking change indicated in the title refers to the removal of the plugin's `doc` command brought about by the upgrade of the SDK version.