chore(website): Make careers page SEO friendly (#13202)
I was recently doing small google search on "CloudQuery careers" and seems it didn't find the right thing. This should help moving out of notion which was nice for getting started but not great for SEO. Also, I think it's quite hard to find it so I also put it on the upper bar. 

Also, move application to typeform from email.