fix: Normalize PKs to not null (#13557)
#### Summary

Saw this while working on https://github.com/cloudquery/cloudquery/issues/13486. If you run migration twice on the same spec it fails. The reason is that MySQL Implicitly sets PKs to non null in the DB, so we have to account for that.

We had this in the code and it seems it got lost in the migration to v3 or v4.
~~I also cleaned up the tests as those are passing without that loop on test options.~~ Tests cleanup moved to https://github.com/cloudquery/cloudquery/pull/13565

Not sure why the double migration tests didn't catch this

<!--