fix!: Split max index length between PKs (#13569)
#### Summary

Fixes https://github.com/cloudquery/cloudquery/issues/13486. The max length of an index is either `3072` bytes or `767` bytes.
Since we use `utf8mb4` which can use up to 4 bytes per char, we max out at either `768` or `191` chars (the number in bytes divided by `4`).
The current code allocates `191` chars per PK which is OK if a table has 4 PKs since `191*4= 764`, but fails if a table has more than that. See list of tables [here](https://github.com/cloudquery/cloudquery/issues/13486#issuecomment-1701366092)

The solution is to get the max number of bytes based on the row format, and evenly distribute the max length between PKs.
This is still not the best solution as we assign the same number of bytes to a UUID and a `text` column (for example CQId is `UUID`), but I don't think it's very common to have one UUID PK and another `text` PK and enough other PKs so the prefix is too short.


<!--