doc: Add multiline text substitution option in the documentation (#13558)
#### Summary

The environment variable substitution documentation is suggesting that new line characters should be escaped for multiline content. This is an unnecessary effort which complicates the scripts to setup CloudQuery environment. 

Users can rather substitute the environment variable using pipe '|' operator and a tabbed new line before the substitution. This is a standard way of importing a multiline text into YAML files.

<!--