feat: Publish command (#13948)
This adds a `cloudquery publish` command to the CLI that can publish plugins to the upcoming CloudQuery registry. It reads a `dist` directory with package.json, created by the plugin-sdk, and calls the CloudQuery API to upload the files and metadata in the directory as a new plugin version. This command will mainly be used by plugin authors to upload a new version of their plugin to the registry.