feat: Add `policy` column to `aws_efs_filesystems` table (#14672)
#### Summary

resolves cloudquery/cloudquery#14632

I'm looking for a way to reduce the duplicate implementations, for example, the following part is a way to get a service to resolve data for the table. It works well when a table relays on a single resolver. But if a table needs multiple resolver then might be a good way to do it. 

```go
	p := resource.Item.(types.FileSystemDescription)
	config := efs.DescribeFileSystemPolicyInput{
		FileSystemId: p.FileSystemId,
	}
	cl := meta.(*client.Client)
	svc := cl.Services(client.AWSServiceEfs).Efs
```

<!--
Explain what problem this PR addresses
-->