chore: Migrate Jira source plugin to premium paid (#15182)
## Removal of Jira source plugin

This PR removes the code of the Jira source plugin from the CloudQuery
monorepo due to lack of usage.

### Why are we doing this?

Development and maintenance of our open source plugins requires a
substantial effort on our side. We have decided that going forward, some
of our plugins will be available as closed-source, premium plugins and
will be available via our [Hub](https://hub.cloudquery.io/).

With this, we are moving the Jira plugin to Premium Preview and its
source code is being removed from our CloudQuery repository (but you can
still access it via the git history).

## How this affects you

If you are using this plugin, you have the following options:

### Keep using the latest release from GitHub

In your config, you can specify `registry: github` and lock on the
latest released version (that was published to GitHub) to keep
installing the plugins from GitHub. To get access to their source code
and create a fork, you can check out their release tags from the
[CloudQuery repository](https://github.com/cloudquery/cloudquery):

- Jira:
[plugins-source-jira-v1.0.10](https://github.com/cloudquery/cloudquery/releases/tag/plugins-source-jira-v1.0.10)

### Log in to Hub and install plugins from there

Premium (free up to a quota) and Premium Preview (free) plugins will be
available upon logging in with CLI. To log in, run the `cloudquery
login` command. You can create a new account at the same time.

To log in in a non-interactive mode, you can [generate an API
key](https://www.cloudquery.io/docs/deployment/generate-api-key).