fix: Change shutdown sequence after cloudquery login (#15448)
Some users have reported an issue where the server takes longer than 5 seconds to shutdown, resulting in them not being logged in. 

This changes the logic so that we always save the token after retrieving it, and only perform the shutdown sequence as the final step (in case it fails). Then at least, even if something goes wrong, they will still be logged in. 

This also extends the timeout from 5 seconds to 15 seconds. One user who reported the issue tested this and the longer timeout worked for them.