chore: Remove old links from docs (#15600)
we had old links. I think now those are not needed and we can just link to home as all the guides are under the blog. I can also add a solutions section just like in our home but it will mean we will need to maintain both which might be a bit annoying.