fix: Add `X-Meta-User-Team-Name` during docker push (#16013)
#### Summary

The requirement for this header was added recently in the backend to track downloads, so the backend needs to know the team that's downloading the plugin. It's less relevant for the push operation though, but still needed. It's ok to use the team from the manifest - it just has to be any team the user is a part of. During push we already verify the token used to push has access to the team.

<!--