chore: Add ddog documentation for OpenTelemetry (#16114)
This improves our OpenTelemetry documentation and adds DataDog example. In general open-telemetry and tracing are advanced topics and it is mostly a developer topic and not really a user topic as the action items are usually for developers to improve scheduling or do code changes. for user we need to provide different information like time for each table as the only thing a user can do is add/remove tables and split the workload better (splitting the workload ideally shouldn't be a thing user is doing as well but we are not there yet with distributed scheduling).