feat: Add table option to filter workflow runs (#17539)
#### Summary

Tried to do https://github.com/cloudquery/cloudquery/issues/17475 using `created_at` column but that's not really possible as while we can filter based on a `created_at` range the results are always returned in descending order, so we always get newest results first.
Which means that if the sync stops in the middle and we only get part of the data, next sync we'd have to split the range to get the runs before the oldest we retrieved and after the newest (and do a split each time we have partial state).

We could maybe save the state only after retrieving all the pages (and not after each page) which should ensure we don't have data to catch up on in the next sync.

This PR does something else, it adds a table option to let users specify the range they'd like to sync.

Opening for feedback