gcs: Support path variables on destination plugin.
This commit follows an approach similar to the S3 plugin.
To maintain backwards compatibility:
Check if the path string has any of the template variables like {{FORMAT}},
{TABLE}} and so on. If they do, we follow the behaviour of S3 plugin wherein we
replace the template variables and that's it. If there are no such variables, we
revert back to the current behaviour in the code block shown above.

This commit is not the complete solution, we need to:
- disable uuid in case of no_rotate and any other possible
schema validations.
- replace path variables in read.go.

We are not doing the above items in this commit for the sake of
breaking down commits in easy to review chunks. Still, this commit
passes all the existing tests to make sure that none of the individual
commits would break any of the existing tests.