feat!: Add path variable support to GCS destination plugin.  (#17602)
Fixes #15097 . Since the PR is fairly large, please review it commit by
commit along with a read of the commit message descriptions to get a
better sense of the changes.

#### Summary

Add path variable support to GCS destination plugin. **This is a
breaking change**. We check if the path string has any of the template
variables like {{FORMAT}}, {{TABLE}} and so on. If they do, we follow
the behaviour of S3 plugin wherein we replace the template variables and
that's it. If there are no such variables, we revert back to the current
behaviour in the code block shown above.

Please read the issue comments to check on how the decisions for this PR
were made.

This PR follows similar approach to the S3 plugin. I've added some extra
tests for client.read() errors which we should probably copy back to S3
and File plugins in a follow up issue.

Note: I have not made additions to CHANGELOG.md, I'm assuming the
maintainer will do it just before merge, in order to make sure of the
right tag at the time of merge. Please lmk if I need to make that change

<details> 
<summary> GCS screenshot for path `new/{{TABLE}}/{{UUID}}.{{FORMAT}}`
</summary>

![Screenshot 2024-04-12 at 1 46
50 AM](https://github.com/cloudquery/cloudquery/assets/13910561/6b17748b-6686-4c42-8e9d-4d40a2c270b3)


</details>

- [x] Read the [contribution
guidelines](https://github.com/cloudquery/cloudquery/blob/main/CONTRIBUTING.md)
🧑‍🎓
- [x] Run `make lint` to ensure the proposed changes follow the coding
style 🚨 (install golangci-lint
[here](https://golangci-lint.run/usage/install/#local-installation))
- [x] Run `make test` to ensure the proposed changes pass the tests 🧪
- [x] If changing a source plugin run `make gen` to ensure docs are up
to date 📝
- [x] Ensure the status checks below are successful ✅

---------

Co-authored-by: Erez Rokah <erezrokah@users.noreply.github.com>