feat: Add custom JSON schema errors (#17669)
#### Summary

Fixes https://github.com/cloudquery/cloudquery-issues/issues/1526 (internal issue).
This PR uses the custom errors syntax from https://ajv.js.org/packages/ajv-errors.html#messages-for-properties-and-items to improve the errors we provide for the S3 destination plugin (and aligns file and GCS plugin with it).
While this won't help with the Go validation until https://github.com/santhosh-tekuri/jsonschema/issues/44 is implemented it does help with our Cloud Syncs, examples below.

The `errorMessage->properties->prop-name` is needed so the UI knows to show the error under the correct field and not as a top level error

Before:
![image](https://github.com/cloudquery/cloudquery/assets/26760571/e5d252a1-5ca5-4650-bde8-ce1b79960eb0)

After:
![image](https://github.com/cloudquery/cloudquery/assets/26760571/3de97312-eaf3-4230-bb3b-cf20aebed0c2)

Before:
![image](https://github.com/cloudquery/cloudquery/assets/26760571/10a0da89-8f1c-4620-adfd-532832915b99)

After:
![image](https://github.com/cloudquery/cloudquery/assets/26760571/2945a600-1fc0-4e08-9ee2-b139c0d90a5d)

(The latter still need some frontend work not to show the top level error)