feat!: Store timestamps as `timestamp_tz` (#17843)
#### Summary

Currently we are storing all timestamps as `timestamp_ntz` this is "wall clock" time meaning that it has no timezone associated with it. the `arrow.TimestampType` is:

```
TimestampType is encoded as a 64-bit signed integer since the UNIX epoch (2017-01-01T00:00:00Z). The zero-value is a second and time zone neutral. Time zone neutral can be considered UTC without having "UTC" as a time zone
```

Which makes it a better fit for users to be able to utilize the data in their queries.