doc: Link to hub from example specs (#18071)
Some customers appear to be using our SEO (export) pages as a
walkthrough and they miss the fact that there's detailed documentation
somewhere else. This is adding a link directly in the example spec to
help users getting back to the detailed documentation directly from the
code. I am using bitly links on purpose to be able to measure whether
this actually had any effect. That is why this change is affecting only
a selected subset of plugins.