fix: Isolate env variables in `test-connection` command if necessary (#18200)
This was missing from the `test-connection` command (we have it in `sync`) and it will be used by the cloud side.
Currently we don't test multiple connections at the same time, but when/if we do, this will come in handy. It also simplifies the handling code as `test-connection` and `sync` behaves the same.