fix: Remove Primary Key Validation (#18234)
#### Summary

Prior to this version BQ would fail if any table has a primary key. Now if a PK is defined it will just ignore it.

All PKs should be removed by the CLI. In the case of state tables, when versioning is enabled, the SDK will handle duplicates by choosing the latest version