fix: Don't use a reader with seeking support when `no_rotate` is true (#18608)
#### Summary

Follow up to https://github.com/cloudquery/cloudquery/pull/18361.
When `no_rotate: true` we can get large files to upload and when that happens we'll be holding the whole file in memory and eventually running out of it.
An example is a sync on many AWS accounts (for example 2000), even for a single table. Let's say the table data is 10MB per account, we end up with a file of 20GB.

The downside of this change is that if a sync has a single account but many tables (and `no_rotate: true`), we'll run into the same issue https://github.com/cloudquery/cloudquery/pull/18361 fixed.

There are workarounds for both issues (either reducing the number of accounts in the sync, or reducing the concurrency).

The best solution would be to get https://github.com/aws/aws-sdk-go-v2/issues/2694 fixed (suggested solution in https://github.com/aws/aws-sdk-go-v2/compare/main...erezrokah:fix/dont_over_allocate_small_files?expand=1), but doesn't seems like it's going to happen soon.