feat: Add logs to Datadog dashboard (#18867)
#### Summary

This adds a new section to our Datadog dashboard with information on throttled APIs:
![image](https://github.com/user-attachments/assets/fda9fcfe-7d3c-4e42-a134-95890e64243a)

Clicking on a column will users open the logs in the relevant location:
![image](https://github.com/user-attachments/assets/e9b931b0-9074-467f-ad4a-b9ec4cd929f5)
![image](https://github.com/user-attachments/assets/06d2c975-df01-4a99-8446-317eb1b56c5d)

I've also added the the same change to https://github.com/DataDog/integrations-extras/pull/2451.

https://github.com/DataDog/integrations-extras/pull/2451 also adds custom log pipelines to extract the retry attempts into attributes instead of relying on the `retrying request` message.
However if we only distribute the dashboard (like we do on our website) we can't ship the pipelines with the dashboard (we need the integration for that).