feat: Implement transformations in migrate cmd. (#19034)
fixes https://github.com/cloudquery/cloudquery-issues/issues/2370

When I implemented transformer plugins, I simply didn't even know that the `migrate` command existed honestly. The reason that the migrate command doesn't apply transformations is because the `migrate` command is an entirely separate command, and only knows about sources & destinations.

This PR is more or less a copy-pasta of the sync_v3 code, which:
- Instantiates transformer plugins
- Upon sending the `MigrateTable` message, applies `TransformSchema` for all existing transformers onto the schema.

Note that there can be multiple destinations, and an unknown heterogeneous topology of transformer pipelines for each (in case the code looks tricky).

I ran some tests that show how `cloudquery migrate .` now applies transformations:
![Screenshot 2024-08-30 at 10 20 16](https://github.com/user-attachments/assets/5d520fb6-4e02-4057-8c42-7e4c19b390a1)
![Screenshot 2024-08-30 at 10 30 26](https://github.com/user-attachments/assets/823c6bba-bd77-49ea-9980-ca7e70441d04)

But when there are no transformers, it continues to work properly:
![Screenshot 2024-08-30 at 10 19 50](https://github.com/user-attachments/assets/a2edaf3b-6086-4480-a831-f76a3f1dfead)
![Screenshot 2024-08-30 at 10 19 32](https://github.com/user-attachments/assets/7d14c1e2-c1ce-42c8-b299-789bf5459044)