fix: Use per OS login command in quick start guide (#19095)
#### Summary

In our quick start guides we tell users to run `cloudquery login`. However we use the same command for all operating systems. `cloudquery login` won't work for Linux (unless added to the PATH) or Windows
![image](https://github.com/user-attachments/assets/76395482-33e9-4d5c-a99e-15594db84c71)

This PR fixes the issue by using `cloudquery login` on macOS, `./cloudquery login` on Linux and `cloudquery.exe login` on Windows. The fix is consistent with how we tell users to run a sync a bit below in the guide:

#### macOS
![image](https://github.com/user-attachments/assets/bd0286e1-f3fd-4d19-9e3f-add6da82800a)
#### Linux
![image](https://github.com/user-attachments/assets/acb2a9ae-dec9-4d3f-8d8d-051bed4b2438)
#### Windows
![image](https://github.com/user-attachments/assets/f2d98353-a4d1-4a2d-9dcb-f7ff773d1250)