refactor: Inline quick start guide configure components (#19100)
#### Summary

At the moment the quick start guide is split into many components making it harder to update it. This was done to support multiple operating systems via components composition. However we can use the same approach as in https://github.com/cloudquery/cloudquery/pull/19095 (a `switch/case` to handle the logic) instead.
That allows us to have a single component with all the guide, making it easier to update