fix(main): Write to mysql in bulk (#19123)
<!-- 🎉 Thank you for making CloudQuery awesome by submitting a PR 🎉 -->

#### Summary

Fixes #19122

Before, the `INSERT` query was built within the `appendTableBatch` and
`overwriteTableBatch` functions. But we aren't yet aware of how many
rows need to be inserted, so we don't know how many value blocks to add.
Rows are iterated within `writeResources` to create the final records to
insert, so this is an opportunity to calculate a row count. So I moved
the query building to within `writeResources` and constructed the
`INSERT` query based on the amount of rows to be inserted.

This reduced the time to write 27,000 records from 20 minutes to under 1
minute.

---------

Co-authored-by: erezrokah <erezrokah@users.noreply.github.com>