fix: Fix lack of error on destination close. (#19178)
@murarustefaan reported a new issue on syncs, where closing a
destination plugin results in the sync continuing to process source
records until the end (but doing nothing with them), and then finishing
with an error. This only happens on certain sync configurations.

The issue is happening because of two reasons:
- A send on a closed pipeline doesn't error (it should):
https://github.com/cloudquery/cloudquery/blob/e151df2d2ee09f4d77f13f8d7e92cb3f092c2b03/cli/internal/transformerpipeline/pipeline.go#L75
- There's still an error when writing to the destination, but errors
propagate forwards (this is expected, as `io.EOF` propagates forwards to
finalise the sync). An error also closes the pipeline, but due to the
first bullet point, it doesn't error the sync.

As a result, when the destination errors, the source happily sends all
records to the ether, and when done, the sync does error.

The fix is just to error when a send is sent to a closed pipeline.

The following video reproduces the error, installs the fix and shows how
it behaves correctly after:


https://github.com/user-attachments/assets/9c7ef3eb-2170-42fd-bf75-6fe300c397a0