fix: Add missing OTEL logs component (#19220)
#### Summary

This fixes an issue with the `--tables-metrics-location` flag, that when you pass it you'd see the following log warnings:

```
2024-09-20T13:28:20Z WRN otel error error="failed to send logs to http://localhost:51875/v1/logs: 404 Not Found" invocation_id=95b765f7-2472-4c7a-b062-c0c799743bf7 module=cli
```

The reason is that now plugins send OTEL logs data as well, but the CLI is not set up to receive them. This was an oversight on my end when adding logs support to plugins.

The PR also cleans up some unused options