chore: Moving to UI generalisation V2 (#19237)
Migrating Postgres destination to UI V2. This follows the pattern established in the MySQL destination plugin.

For Postgres, we've tried to preserve the usage of key-value pair connection strings when using the regular setup. This allows special characters to be used without escaping (due to the value being encompassed in `'`), in addition it allows the UI to be easily re-populated when the plugin is edited.

refs: https://github.com/cloudquery/cloudquery-issues/issues/2474