chore(main): Updated docs footer and header to match marketing site (#19280)
I updated the footer and the header in the documentation to better align
with the current state of our navigation on the marketing website. The
best practice is to have them be the same. I also updated the logo on
the docs in the upper left-hand corner to redirect back to the homepage,
which is also the best practice in the industry.

In the future, I would like to have the NAV bar be the same across the
docs and our marketing website, but this will work for now.

New Footer:
<img width="1260" alt="image"
src="https://github.com/user-attachments/assets/d32d2e42-2e05-4375-846f-b6cb3e816ea4">

New Header:
<img width="1050" alt="image"
src="https://github.com/user-attachments/assets/1e5450f1-ec5c-42af-aab7-b4e6863bc833">