fix: Send content type when uploading images (#19373)
#### Summary

Fixes https://github.com/cloudquery/cloudquery-issues/issues/2567 (internal issues). We need to update the CLI to adjust for a change in our backend requiring a content type when uploading images. This happens in the CLI when publishing plugins or addons docs that reference local images in their markdown doc files.