fix: Return correct error from transformers pipeline (#19448)
#### Summary

Fixes https://github.com/cloudquery/cloudquery-issues/issues/2591 (internal issue)

We have 2 Go routines in play in `sync_v3`, one for the transformer pipeline and one for handling sync messages on the gRPC channel.
Any of those Go routines can return an error, and the error group will report the first one which is not consistent in case the pipelined was closed due to a destination error where `OnOutput` returns the error and closes the pipeline.
In the latter case we should signal the sync Go routine to exit early but not with an error that will get reported from the pipeline Go routine.